chore: splits games descriptors from setup facilities
