+++ apps/games/assets/coinche/index.js
@@ -1,0 +1,10 @@
+export * from './logic/build.js'
+
+/**
+ * Localized metadatas
+ */
+export const locales = {
+  fr: {
+    title: 'Coinche'
+  }
+}
