+++ apps/games/assets/klondike/index.js
@@ -1,0 +1,20 @@
+export * from './logic/build.js'
+
+/**
+ * Localized metadatas
+ */
+export const locales = {
+  fr: {
+    title: 'Solitaire'
+  }
+}
+
+/**
+ * Minimum time in minutes.
+ */
+export const minTime = 15
+
+/**
+ * Minimum age in years.
+ */
+export const minAge = 7
