--- apps/server/tests/fixtures/games/belote.js
@@ -1,3 +1,0 @@
-export const meshes = [{ shape: 'card', id: 'one-of-diamonds' }]
-export const bags = new Map()
-export const slots = []
