+++ apps/server/tests/fixtures/games/belote/index.js
@@ -1,0 +1,7 @@
+export function build() {
+  return {
+    meshes: [{ shape: 'card', id: 'one-of-diamonds' }],
+    bags: new Map(),
+    slots: []
+  }
+}
