--- apps/server/tests/fixtures/games/klondike.js
@@ -1,3 +1,0 @@
-export const meshes = []
-export const bags = new Map()
-export const slots = []
