+++ apps/server/tests/fixtures/games/klondike/index.js
@@ -1,0 +1,7 @@
+export function build() {
+  return {
+    meshes: [],
+    bags: new Map(),
+    slots: []
+  }
+}
