--- apps/server/tests/fixtures/games/splendor.js
@@ -1,5 +1,0 @@
-export const restricted = true
-export const rulesBookPageCount = 4
-export const meshes = []
-export const bags = new Map()
-export const slots = []
