+++ apps/server/tests/fixtures/games/splendor/index.js
@@ -1,0 +1,9 @@
+export const restricted = true
+export const rulesBookPageCount = 4
+export function build() {
+  return {
+    meshes: [],
+    bags: new Map(),
+    slots: []
+  }
+}
