--- apps/server/tests/fixtures/invalid-games/invalid.js
@@ -1,9 +1,0 @@
-export const cards = [
-  { id: '1' },
-  { id: '2' },
-  { id: '3' },
-  { id: '4' },
-  { id: '5' }
-]
-// slots should be an array
-export const slots = 100
