+++ apps/server/tests/fixtures/invalid-games/invalid/index.js
@@ -1,0 +1,7 @@
+export function build() {
+  return {
+    mesh: [{ id: '1' }, { id: '2' }, { id: '3' }, { id: '4' }, { id: '5' }],
+    // slots should be an array
+    slots: 100
+  }
+}
